TS-620 And even more 64-bit fixes to the SM
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1056194 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cache/CacheTest.cc
@@ -213,8 +213,10 @@ void CacheTestSM::fill_buffer() {
     int64_t l = avail;
     if (l > sk)
       l = sk;
+
     int64_t pos = cvio->ndone +  buffer_reader->read_avail();
-    int o = pos % sk;
+    int64_t o = pos % sk;
+
     if (l > sk - o)
       l = sk - o;
     k.b[0] = pos / sk;
@@ -236,7 +238,7 @@ int CacheTestSM::check_buffer() {
     int64_t l = avail;
     if (l > sk)
       l = sk;
-    int o = pos % sk;
+    int64_t o = pos % sk;
     if (l > sk - o)
       l = sk - o;
     k.b[0] = pos / sk;
