+++ iocore/cluster/P_ClusterInternal.h
@@ -494,14 +494,14 @@ extern void cluster_raise_priority(ClusterHandler *, ClusterVConnState *);
 extern void cluster_schedule(ClusterHandler *, ClusterVConnection *, ClusterVConnState *);
 extern void cluster_reschedule(ClusterHandler *, ClusterVConnection *, ClusterVConnState *);
 extern void cluster_disable(ClusterHandler *, ClusterVConnection *, ClusterVConnState *);
-extern void cluster_update_priority(ClusterHandler *, ClusterVConnection *, ClusterVConnState *, int, int);
+extern void cluster_update_priority(ClusterHandler *, ClusterVConnection *, ClusterVConnState *, int64_t, int64_t);
 #define CLUSTER_BUMP_NO_REMOVE    -1
 extern void cluster_bump(ClusterHandler *, ClusterVConnectionBase *, ClusterVConnState *, int);
 extern int iov_memcpy(IOVec *, int, int, char *);
 
 extern IOBufferBlock *clone_IOBufferBlockList(IOBufferBlock *, int, int, IOBufferBlock **);
-extern IOBufferBlock *consume_IOBufferBlockList(IOBufferBlock *, int);
-extern int bytes_IOBufferBlockList(IOBufferBlock *, int);
+extern IOBufferBlock *consume_IOBufferBlockList(IOBufferBlock *, int64_t);
+extern int64_t bytes_IOBufferBlockList(IOBufferBlock *, int64_t);
 
 // ClusterVConnection declarations
 extern void clusterVCAllocator_free(ClusterVConnection * vc);
